move ConnWaitGroup to testing package
