+++ client_test.go
@@ -6,11 +6,11 @@ import (
 	"io"
 	"math/rand"
 	"net"
-	"sync"
 	"testing"
 	"time"
 
 	"github.com/segmentio/kafka-go/compress"
+	ktesting "github.com/segmentio/kafka-go/testing"
 )
 
 func newLocalClientAndTopic() (*Client, string, func()) {
@@ -72,8 +72,8 @@ func newLocalClient() (*Client, func()) {
 }
 
 func newClient(addr net.Addr) (*Client, func()) {
-	conns := &connWaitGroup{
-		dial: (&net.Dialer{}).DialContext,
+	conns := &ktesting.ConnWaitGroup{
+		DialFunc: (&net.Dialer{}).DialContext,
 	}
 
 	transport := &Transport{
@@ -88,31 +88,6 @@ func newClient(addr net.Addr) (*Client, func()) {
 	}
 
 	return client, func() { transport.CloseIdleConnections(); conns.Wait() }
-}
-
-type connWaitGroup struct {
-	dial func(context.Context, string, string) (net.Conn, error)
-	sync.WaitGroup
-}
-
-func (g *connWaitGroup) Dial(ctx context.Context, network, address string) (net.Conn, error) {
-	c, err := g.dial(ctx, network, address)
-	if err != nil {
-		return nil, err
-	}
-	g.Add(1)
-	return &groupConn{Conn: c, group: g}, nil
-}
-
-type groupConn struct {
-	net.Conn
-	group *connWaitGroup
-	once  sync.Once
-}
-
-func (c *groupConn) Close() error {
-	defer c.once.Do(c.group.Done)
-	return c.Conn.Close()
 }
 
 func TestClient(t *testing.T) {
